chore: split theme type declarations
